add responsive navbar and refactor card component
