+++ components/general/LatestNews.tsx
@@ -1,8 +1,8 @@
 import { fontSyne } from "@/pages/_app";
 import styles from "@/styles/LatestNews.module.css";
-import Card from "../base/Card";
 import Button from "../base/Button";
 import Divider from "../base/Divider";
+import Card from "../base/Card";
 
 const LatestNews = () => {
   return (
@@ -19,21 +19,25 @@ const LatestNews = () => {
       </div>
       <article className={styles.latest_news__article}>
         <Card
-          image={"/images/two_people_praying.png"}
-          cornerMakerLeft={"Branding, Design"}
-          title={"Things to Look for When Comparing Branding Alternatives"}
+          title="Things to Look for When Comparing Branding Alternatives"
+          subtitle="Branding, Design"
+          image="/images/two_people_praying.png"
+          direction="column_reverse"
+        />
+
+        <Divider />
+        <Card
+          title="5 Stand-out Features of Branding You Should Know"
+          subtitle="Branding, Design"
+          image="/images/channel.png"
+          direction="column_reverse"
         />
         <Divider />
         <Card
-          image={"/images/channel.png"}
-          cornerMakerLeft={"Branding, Design"}
-          title={"5 Stand-out Features of Branding You Should Know"}
-        />
-        <Divider />
-        <Card
-          image={"/images/brand_strategy.png"}
-          cornerMakerLeft={"Branding, Design"}
-          title={"Branding: What Real Customers Have to Say"}
+          title="Branding: What Real Customers Have to Say"
+          subtitle="Branding, Design"
+          image="/images/brand_strategy.png"
+          direction="column_reverse"
         />
       </article>
     </section>
