+++ styles/Card.module.css
@@ -1,36 +1,44 @@
 .card {
-  width: 530px;
   display: flex;
   flex-direction: column;
+  flex: 1;
   gap: 16px;
 }
 
 .card__image {
-  background-color: #7d7d7d;
-  border-radius: 16px;
-  height: 100%;
+  border-radius: 12px;
   object-fit: cover;
-  display: block;
 }
 
-.card__text {
-  height: 107px;
+.card__title_container_column {
   display: flex;
   flex-direction: column;
-  row-gap: 8px;
 }
 
-.card__text__corner {
-  font-style: normal;
-  font-weight: 400;
-  text-transform: uppercase;
-  color: #191919;
+.card__title_container_row {
+  display: flex;
+  flex-direction: row;
+  justify-content: space-between;
+  align-items: center;
 }
 
-.card__text__title {
-  width: 530px;
-  height: 84px;
-  font-style: normal;
-  font-weight: 500;
-  color: #191919;
+.card__title_container_column_reverse {
+  display: flex;
+  flex-direction: column-reverse;
+  gap: 8px;
+}
+
+.card__image_large {
+  width: 100%;
+}
+
+.card__image_medium {
+  width: 100%;
+  height: auto;
+  min-height: 760px;
+}
+
+.card__image_small {
+  width: 100%;
+  height: 380px;
 }
