+++ styles/Hero.module.css
@@ -26,6 +26,7 @@
 .hero__photo {
   border-radius: 16px;
   height: 100%;
+  width: 100%;
 }
 
 .hero__icon {
