+++ styles/LatestNews.module.css
@@ -24,4 +24,5 @@
   display: flex;
   flex-direction: row;
   justify-content: space-between;
+  gap: 48px;
 }
