--- styles/MemberCard.module.css
@@ -1,44 +1,0 @@
-.card {
-  display: flex;
-  flex-direction: column;
-  flex: 1;
-  gap: 16px;
-}
-
-.card__image {
-  border-radius: 12px;
-  object-fit: cover;
-}
-
-.card__title_container_column {
-  display: flex;
-  flex-direction: column;
-}
-
-.card__title_container_row {
-  display: flex;
-  flex-direction: row;
-  justify-content: space-between;
-  align-items: center;
-}
-
-.card__title_container_column_reverse {
-  display: flex;
-  flex-direction: column-reverse;
-  gap: 8px;
-}
-
-.card__image_large {
-  width: 100%;
-}
-
-.card__image_medium {
-  width: 100%;
-  height: auto;
-  min-height: 760px;
-}
-
-.card__image_small {
-  width: 100%;
-  height: 380px;
-}
